pyproject.toml
[tool.poetry]
name = "task-management"
version = "0.1.0"
description = ""
authors = ["nadduli <naddulidaniel94@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.32.3"
pyjwt = "^2.9.0"
pycodestyle = "^2.12.1"
fastapi = {extras = ["standard"], version = "^0.115.4"}
sqlmodel = "^0.0.22"
passlib = "^1.7.4"
aioredis = "^2.0.1"
pydantic-settings = "^2.6.1"
asyncpg = "^0.30.0"
autopep8 = "^2.3.1"
click = ">=8.0, <9.0"
black = ">=23.0, <24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
